- Move OS information functions to util.rb
